Refactor from `dao` to `repository`
