+++ src/main/java/com/kerrlabajo/tour_of_heroes/repository/HeroDao.java
@@ -1,4 +1,4 @@
-package com.kerrlabajo.tour_of_heroes.dao;
+package com.kerrlabajo.tour_of_heroes.repository;
 
 import com.kerrlabajo.tour_of_heroes.model.Hero;
